Changed Camera System, Near Seemless Zoom to Solar View
Changed the camera system from a moving one a still one with the object
moving around it. This allowed me to use one coordinate system for the
entire galaxy with doubles and zoom in seemlessly from the galaxy view
to the solar view, allowing for more imersive and intuitive gameplay.