Re-added Companies and Ships; Worked on Info Panel
Readed the near bare company and ship scripts, with updates to get them
current.
Majority of the work was done reorganizing the storage and saving of
information like items, ships, structures, and creatures. With the newly
implemented blueprint system adding extra items into the game to be used
by ai should be a lot easier.
Also worked on a proper and unified info panel to be used to display
contextual information about game entities like creatures, governments,
solar bodies, etc.