Semi-working Supply and Demand
I wanted the game to automatically set the prices of each item based on
the supply and demand of the game intities without the need to set a
base price. This should allow for a more dynamic game with interesting
occurrences. That is partially done, need to work on how it will
recognize supply.

Items created for some basic materials that allow for the creation of
drills and factories. A special structure called the buildstructure is
used to designate a demand for an item on a SolarBody, this will then
increase the demand for said item in the marketplace until it is high
enough for factories to supply it at profit.

More info panel configurations and information increased.