Start of Better Marketplace
Initial goap implenetation done, but the buying and selling of goods
isn't good enough for me to test GOAP on it. I started to fix it by
reorganizing the production structure classes and adding a temporary
planet needs for supplying the planet through the space station on the
populated planet (Better implementation will be needed in the future).

TODO: Better company/government management of owned structures and item
needs. Better management of buying and selling of goods.