Drillers and Factories Connected
Drillers and factories of the same company/government now connect to
their suppling structures as well as their output structures. This
connection takes into account the needed rate of the item produces, and
sends the extra to ground storage if there is any extra.

Factories are now setup to run not just based on profit/cost but also if
their ouput factories are stocked, allowing for a better run industry.