Update to Unity 2018 beta
Updated to Unity 2018 beta and created a build to showcase what I have
done so far. Will be doing some drastic changes to the game after this.
Ex. converting back to float and breaking the distances back into steps.
Creating a localized view for the displaying of 3d ships, stations, and
planets, and creating a multi camera setup to use for the local and
galaxy maps.